lib/requirements.txt
PyYAML==6.0.1
requests==2.32.3
jsonschema==4.23.0
rapidfuzz==3.10.1
